Discard stdout, instead of sending it to clients
Language servers implemented with lsp4clj tend to send their output over
stdout. Clients read this output and parse it as JSON-RPC. So, it
shouldn't contain anything besides what the server intended to send.

But, if any server prints to stdout, even accidentally, that output will
be intermingled with the JSON-RPC messages.
https://github.com/clojure-lsp/clojure-lsp/issues/1080 was one such
issue. Interestingly Calva users experienced this issue, but Emacs users
didn't, suggesting that some clients handle unexpected data in the
server output better than others.

This patch adds a helper that captures stdout and discards it, instead of
sending it to clients.

It protects all of lsp4clj's uses of ILSPFeatureHandler and
ILSPProducer. This includes:
- Uses of the feature handler and producer in LSPServer.
- Uses of the feature handler and producer in LSPTextDocumentService.
- Uses of the feature handler and producer in LSPWorkspaceService.

It does not protect any of the following:
- Server code that runs before `initialize`, or any async processes
  started during this time. (e.g. cloure-lsp.server/run-server! and
  spawn-async-tasks!)
- Any code that wraps ILSPFeatureHandler or ILSPProducer (e.g.
  ClojureLSPFeatureHandler and ClojureLspProducer)
- Any extensions of the Server or Services (e.g. ClojureLanguageServer)
- Any code that uses the wrappers or extensions (e.g. ClojureLspServer)

Servers should make efforts to protect against accidental output to
stdout in all of these scenarios, by using the helper or a similar one.