Try different approach to shutdown
Clients have two ways to shutdown. They can send the 'shutdown' request,
or they can close their end of the server input. The shutdown is mostly
for the language server to do cleanup, so now we do nothing. On exit, we
simulate the server input closing, which blocks waits for messages to
finish being processed, then closes the output, then unblocks.