Support communication through a socket
Adds lsp4clj.server/socket-server, as an alternative to
lsp4clj.server/stdio-server.

```clojure
(lsp4clj.server/socket-server {:port 61235})
```

This starts a socket server listening on the provided port, blocking
until a client makes a connection. When the connection is made it
returns a lsp4clj server that has the same behavior as a `stdio-server`,
except that messages are exchanged over the socket. When the server is
shut down, the connection will be closed.

Fixes #1