Improve helpers for creating requests and responses
This re-models the request and response helpers after
ring.util.response. The new helpers either create request or response
objects, or assist in updating them.

Note that this deletes the json-rpc.messages namespace without
deprecating it, which is a breaking change. However, I don't expect that
there were many users of that namespace (clojure-lsp uses it only in
integration tests).