Allow servers to place requests while receiving
Messages are read off the input channel one at a time. In the case of
requests, the pipeline is blocked until the request completes. But this
causes a bug. During an inbound request a server can place an outbound
request. Because the pipeline is blocked, the client's response can't be
received until the first request finishes. If the server blocks waiting
for a response without a timeout, we end up deadlocked. And even if the
server uses a timeout, it doesn't receive the response in time to use
it.

This splits the input stream in two. The inbound messages that are
initiated by the client (inbound requests and notifications) are
processed in one stream, and inbound messages initiated by the server
(inbound responses) are processed in another. This allows a server to
receive a response even while it's waiting to finish processing a
request.

This commit doesn't change the number of threads used. Before, when we
were using pipeline-blocking, one thread was created to processes all
inbound messages. Now one thread processes all client-initiated inbound
messages. The rest of the input channel message negotiation is handled
in go blocks.

This thread is necessary so that we can block writes to the output
channel, to respect backpressure from clients that are slow to read.