Process requests in parallel
This is an attempt to fix
https://github.com/clojure-lsp/clojure-lsp/issues/1240, where the
failure to run requests in parallel leads to small freezes while typing
in Emacs. In other editors it manifests differently. In Calva, we've
noticed that diagnostics and semantic tokens are delivered more slowly.

This commit changes lsp4clj to allow language servers to process and
respond to up to 4 requests in parallel. This is a departure from the
[spec](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#messageOrdering),
which says "Responses to requests should be sent in roughly the same
order as the requests appear". However, it is closer to what lsp4j and
cloure-lsp were doing before the migration away from lsp4j.

In the original implementation, lsp4j
[consumed](https://github.com/eclipse/lsp4j/blob/04b0c6112f0a94140e22b8b15bb5a90d5a0ed851/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/RemoteEndpoint.java#L185-L196)
messages one at a time, handing them off to the language server. The
language server processed each message serially.

lsp4clj fully [processed each
notification](https://github.com/eclipse/lsp4j/blob/04b0c6112f0a94140e22b8b15bb5a90d5a0ed851/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/RemoteEndpoint.java#L220)
before it moved on to the next message.

But requests were a little different. Like notifications, requests were
handed to the language server in series. But the language server
[returned](https://github.com/eclipse/lsp4j/blob/04b0c6112f0a94140e22b8b15bb5a90d5a0ed851/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/RemoteEndpoint.java#L261)
a CompletableFuture which could, at the server's option, be incomplete.
lsp4j would
[arrange](https://github.com/eclipse/lsp4j/blob/04b0c6112f0a94140e22b8b15bb5a90d5a0ed851/org.eclipse.lsp4j.jsonrpc/src/main/java/org/eclipse/lsp4j/jsonrpc/RemoteEndpoint.java#L279)
for the result of that future, whenever it was done being calculated, to
be sent back to the client as the response. This let the language server
decide how much of the response to calculate (if any) while blocking
other requests and notifications, and how much to calculate in another
thread, allowing other requests and notifications to be processed in
parallel.

If the response creation was wrapped in `CompletableFuture/supplyAsync`,
as was the case for [almost
all](https://github.com/clojure-lsp/lsp4clj/blob/fe8782941fd871921d55349b9f7803ac0dc8595d/server/src/lsp4clj/core.clj#L100)
of clojure-lsp's requests, then the response would be calculated in the
"common" ForkJoin thread pool. This thread pool has a configurable
parallelism (on my machine it's 3). The end result is that clojure-lsp
processed most of its requests in parallel—up to 3 at a time.

I didn't understand this arrangment as I was migrating lsp4clj away from
lsp4j. I thought that lsp4j procssed every request in series, so that it
could follow the spec's guidance. I followed that pattern, making all
the requests and notifications serial. (Actually I started with an
implementation that distributed both requests and notifications to the
server in parallel. That meant the server processed requests in parallel
but also that it received message out of order. That parallelism was
removed in 6996fdbaefe743431994839db07676a24052cd25.)

Technically, the strictly serial implementation works, but it causes the
lagginess and delays observed in #1240. So, this patch is an attempt to
restore the earlier behavior, though with a different implementation.

In this implementation lsp4clj continues to read notifications and
requests in series, and passes them to the language server with
lsp4clj.server/receive-notification and lsp4clj.server/receive-request.
The bodies of these functions are executed serially. Before this commit
(though after the migration), the return value of receive-request was
used as the response, which meant that each response was fully
calculated serially. But now the language server can, optionally, return
a deref-able object from receive-request. If it does, the deref-able
will be derefed in a thread pool of size 4. So, a language server can
indicate to lsp4clj that it would like a request to be processed in
parallel with other requests by, for example, returning a `delay`. Of
course, if it wants a request to block other requests, it can still do
that by calculating the full response synchronously in receive-request.
Or it can do some of both, calculating part of a response, then
wrapping the rest of the calculation in a delay.

There were four places where clojure-lsp processed responses
synchronously (by using `CompletableFuture/completedFuture` instead of
`CompletableFuture/supplyAsync`):
`initialize`, `shutdown`, `rangeFormatting` and `executeCommand`.

I think it's important that clojure-lsp continues to process the
`initialize` request before it allows others through. But the others
could be re-considered someday:

* `executeCommand`. On the clojure-lsp side, `executeCommand` starts a
  `future` and returns nil. The only difference that I see from using
  `supplyAsync` (or now `delay`) is that the client sees a `nil`
  response before the command is finished executing. If this isn't
  important—if the client would be content with the command executing
  asynchronously, but getting a nil only after it had finished—then
  `executeCommand` could use `delay`.
* `rangeFormatting`.  On the clojure-lsp side, rangeFormatting is an
  anomaly. It calculates its response synchronously, at the exclusion of
  other requests. On top of this, unlike every other request, it sets
  and releases a global lock, presumably as a way to double-check that
  if multiple rangeFormatting requests come in, the later ones are
  dropped. I'm not sure why this happens—it seems like a bug fix from
  long ago, though I haven't tracked down which bug it was. It seems odd
  though—by blocking other requests, there can't be other simultaneous
  rangeFormatting requests, so isn't the lock useless? Anyway, I don't
  currently see any reason why this couldn't be asynchronous (especially
  since it already uses `process-after-changes`) though it would need
  more research.
* `shutdown`. The clients aren't likely to send requests or
  notifications after sending `shutdown`, so I don't see a good reason
  for it to be synchronous, though it probably doesn't hurt.

I've tested this with clojure-lsp, wrapping most of its requests in
`delay`s. This does fix #1240, so I think it should be merged. However,
in the long run I'm not sure mimicking the old behavior is an ideal
solution. Part of the reason we're seeing these delays is because we're
processing every request fully, even if it's become out of date. The
client is capable of indicating that it no longer needs a response by
sending a $/cancelRequest, but we're ignoring those. We also don't stop
an in-progress analysis if we've learned that the text of a file has
changed since the analysis was started. Fixing that kind of thing might
make these spec-bending compromises less necessary.